HADOOP-13786 patch 013.
All mock tests are working. More specifically, they were got working, then the commit logic tuned to reduce the number of S3 calls (no exists() check before a delete() if not needed, etc),  and the tests tuned to be a bit more explicit about what they are declaring and asserting. Everything is working except  the IT protocol tests testMapFileOutputCommitter and testConcurrentCommitTaskWithSubDir, which expect directories to be handled. I will do that at least for the Directory Committer, with some mock tests as well as a fixed IT test.

+ some tuning of LambdaTestUtils; Ryan's patch had some assertThrown tests which can be moved to a common test base, but while LTU had some features the patch lacked, LTU didn't support handing down extra diagnostics messages. Fixed. We could go one step further and allow callers to provide a closure () -> String for diagnostics, perhaps, though maybe we can wait to see what JUnit 5 has first